roles done

+++ routers/logout.php
@@ -1,18 +1,18 @@
 <?php
     function route($method, $urlList, $requestData){
         include_once './helpers/headers.php';
+        include_once './helpers/checks.php';
 
         global $Link;
 
         if ($method == 'POST'){
             $token = substr(getallheaders()['Authorization'], 7);
-            if (!empty($token)){
+            if (!empty($token) && checkIfTokenIsExist($token)){
                 $logoutResult = $Link->query("DELETE FROM tokens WHERE value='$token'");
                 if ($logoutResult){
                     echo "success logout";
                 }
                 else {
-                    // echo $Link->errno . " - " . $Link->error ;
                     setHTTPStatus("500", "Some troubles with logout'");
                 }
             }
